Added Beeline agent method to get shortest path
Back from its graph using networkx built in method. Reversed the
turns/actions and reversed direction of the graph so we can graph it
and make sure it's right

Removed id from WumpusNode as it didn't make sense.

Added helper methods in Orientation and Action to help out with the
above.

Moved display_graph method to WumpusDiGraph