Fixed issue with incorrect Agent orientation init
In Environments init, the agent was initialized without specifying east
orientation, counting on default argument to take care of that. During
unit testing it was clear that the orientation of the agent in the
previous environment affected the new environment agent orientation.

Found out issue was due to Agent default parameter being Orientation
which is a mutable class (https://stackoverflow.com/questions/1132941/least-astonishment-and-the-mutable-default-argument)

Switched to using OrientationState and then instantiating new
Orientation within Agent init method