Completed Assignment 2, BeelineAgent implemented
BeelineAgent finds the quickest/safest path back once it's found the
gold. Used `Networkx` to keep track of the nodes/locations and
actons/edges that cause a move or orientation change. The agent is able
to figure out the shortest path back even if adjacent locations weren't
visited by the agent consecutively.