ProbAgent methods for building up bayesian network
For both pits/breezes, and wumpus/stenches. Also wrote methodto get the
update probability distribution when evidence is presented to the
network (i.e. percept or similar).

Test method written for the testing of the update of the probability
distribution. Although still haven't designed the method to update the
distribtion in the original model..

Also added adjacent_cells method in the ProbAgent class rather than
keeping it just in the Environment class